WebApp/requirements.txt
streamlit==0.57.3
tensorflow==2.0.0
numpy==1.17.0
Pillow==7.1.1
